backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
pymongo==4.6.1
python-dotenv==1.0.0
Werkzeug==3.0.1
bcrypt==4.1.2
opencv-python==4.9.0.80
numpy==1.26.4
